Add human friendly word list
This list is inspired by but distinct from the list used in many wordle
games designed to be played by humans. It can be used to approximate
how well a bot would perform in such games. It has been tuned to be
ever so slightly harder than the original using a greedy max-entropy
strategy.

The benefit of a new word list is that it is easily available using
an API call on this package.

Support for split word lists is attempted but only tested for the new
fire based CLI.

Also change how word lists are loaded to make it equally easy to load
one of the bundled word lists or a custom one. The new format
optionally contains an empty line to separate guesses from answers.
Advantages of this include
* compact (much smaller than dumping the dict as json),
* compatible with existing word lists

I would have added the files using lfs to prevent bloating the repo
but that is not possible with public forks.